initial implementation
The process works but it definetely needs some polishing. The container
contains only the SOGo. The memcached and the HTTP server runs on the
host as well as the cron jobs which exec into the container.

All the host configuration is handle via Puppet which is - currently -
stored in a different repository.